rendering: Seperate KaTeX specific rendering part
This will allow easier integration of alternative latex to HTML/MathML
backends.
This commit does not change any functionality.
